+++ src/components/routes/blog/index.ts
@@ -1,2 +1,3 @@
 export { BlogPostPage } from "./BlogPostPage";
 export { BlogIndexPage, BlogLayout, BlogLoadingState } from "./BlogIndexPage";
+export { BlogPagination } from "./BlogPagination";
